Add PhysicsBodyComponent and simulate it using dyn4j

+++ core/src/ru/mofr/popballs/components/PositionComponent.java
@@ -3,6 +3,6 @@ package ru.mofr.popballs.components;
 import com.artemis.Component;
 
 public class PositionComponent extends Component {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 }
